Parse: do prefix check before len check (#1)
In practice, it's far more common for bad IDs to simply be of the wrong
"type" than for the UUID part to be mangled. This PR has Parse do the
prefix check first, that way users in practice get much more helpful
errors about having the wrong kind of ID than about having the wrong
length.